Bump dependencies and remove server crypto
On the one hand, I really liked the crypto stuff. However, on the other
hand, I simply don't want to trust crypto performed for me on another
machine. Thus, i removed the code which introduced lots of complexity.

The code will be available in the server-crypto branch. Thanks for all
the work, Markus.