Client side caching with HTTP 304 (Not Modified)
Each served item contains a bogus `Last-Modified` HTTP response header
which a browser can be sent back in a subsequent request as the
`If-Modified-Since` header. The real item creation date is not being
used due to privacy concerns.

If gosh's internal HTTP handler detects a request with an
`If-Modified-Since` header with a date between an item's creation and
expiration, a simple HTTP status code 304, Not Modified, without the
content will be returned.

Closes #8.