Uncouple Store and Item
Prior to this change, some IO operations on Items were part of the Item
type, not the Store. However, there was a coupling between those two as
the Store has passed it's data directory.

With this change, the Store's API became "simpler" in terms of types,
e.g., the ID will be passed as a string instead of an Item, and takes
care of all IO operations. Those are preparations for a privilege
separated store daemon with some network resp. IPC in between.