Store RPC: initial state
The Store RPC is being designed as a "simple" RPC with both a server and
a client part for remote Store access. Remote is thereby defined as a
Unix domain socket or - more precisely a socketpair(2) - which will be
used between two processes for the privilege separation.

So far, its construct and a ping method for testing was implemented and
is being tested in a Go test.

As a drive-by change, the API of Store.Get was altered to drop the
delExpired flag as it was merged with the NewStore's cleanup flag.