Drop user privileges to configured user and chroot
After each subprocess has finished their setup phase, the POSIX
privileges are being dropped to a minimum. This includes a chroot to
either the store's path or a just created bottomless pit (empty root
jail directory) and a change of the effective user and group to the
configured pair from the configuration.