Rework NixOS module
This is a draft.

This PR changes the provided NixOS module to adjust to recent changes to
the codebase. It fails to build at hardening steps. Since this is not my
area of expertise I'm happy to ask for help.

TODO
- Fix hardening
- Implement MIME mapping
- Implement custom_index and static_files paths
- Implement url_prefix

Please note: This commit would also change src from locally to
fetchFromGitHub; please adjust accordingly if this ain't wanted.